backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
celery[redis]
SQLAlchemy>=2.0
alembic>=1.12
httpx
openai>=1.0
google-generativeai>=0.3.0
python-dotenv>=1.0